BSIS-2923 Add suppress warning to BaseViewModel equals method
It can be suppressed safely as the class is being checked before.

Also removed an unnecessary casting in BloodTestingRuleEngineTest.

+++ src/test/java/org/jembi/bsis/service/BloodTestingRuleEngineTest.java
@@ -256,7 +256,7 @@ public class BloodTestingRuleEngineTest extends ContextDependentTestSuite {
       BloodTestResultViewModel bloodTestResultViewModel = resultViewModelMap.get(key);
       if (bloodTestResultViewModel.getReEntryRequired().equals(true)
           && bloodTestResultViewModel.getBloodTest().getBloodTestType().equals(BloodTestType.BASIC_TTI)) {
-        reEntryRequiredTTITestIds.add((Long) bloodTestResultViewModel.getBloodTest().getId());
+        reEntryRequiredTTITestIds.add(bloodTestResultViewModel.getBloodTest().getId());
       }
     }
     Assert.assertEquals("Re-entry required TTI tests", 1, reEntryRequiredTTITestIds.size());
